made installation script more robust in case of relative paths and wrong fastjet path

+++ SFrameInstall.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 if [ "$#" -ne 2 ] ; then
-    echo "Usage: source SFrameInstall.sh <directory to install sframe> <directory of fastjet lib directory>"
+    echo "Usage: ./SFrameInstall.sh <directory to install sframe> <directory of fastjet lib directory>"
     return
 fi
 
@@ -10,18 +10,21 @@ if [ "${ROOTSYS}" = "" ] ; then
     return 
 fi
 
-export SFRAMEDIR=$1
-export FASTJETDIR=$2
+SFRAMEDIR=`readlink -f $1`
+FASTJETDIR=`readlink -f $2`
 
 if [[ -e $SFRAMEDIR ]]; then
     echo "Warning: directory $SFRAMEDIR already exists, content may be overwritten"
 fi
 
+if [ ! -e $FASTJETDIR/libfastjet.so ]; then
+    echo "Error: fastjet dir does not contain 'libfastjet.so'";
+    return 1;
+fi
+
 svn co https://sframe.svn.sourceforge.net/svnroot/sframe/SFrame/tags/SFrame-03-06-11 $SFRAMEDIR
 
 cd $SFRAMEDIR
-touch fullsetup.sh
-
 echo 'export FASTJETDIR='${FASTJETDIR}' \nexport LD_LIBRARY_PATH="'$FASTJETDIR:${SFRAMEDIR}'/SFrameTools/JetMETObjects/lib:$LD_LIBRARY_PATH" \nsource setup.sh' > fullsetup.sh
 source fullsetup.sh
 
@@ -32,7 +35,7 @@ git clone https://github.com/UHHAnalysis/SFrameTools.git SFrameTools
 git clone https://github.com/UHHAnalysis/SFrameAnalysis.git SFrameAnalysis
 
 cd NtupleWriter
-source configure.sh
+./configure.sh
 make -j 8
 cd ../SFrameTools
 make -j 8
@@ -40,3 +43,4 @@ cd JetMETObjects
 make -j 8
 cd ../../SFrameAnalysis
 make -j 8
+
